Split test_utils part1
This is to make test builds & ci faster.
