+++ src/test/c_api_tests.cpp
@@ -44,6 +44,9 @@
 #include "c_api_test_utils.hpp"
 #include "mockmodelinstancechangingstates.hpp"
 #include "test_utils.hpp"
+#include "light_test_utils.hpp"
+#include "platform_utils.hpp"
+#include "test_with_temp_dir.hpp"
 
 using namespace ovms;
 using testing::ElementsAreArray;
