+++ src/test/disabled_mediapipe_test.cpp
@@ -23,6 +23,7 @@
 #include "../server.hpp"
 #include "test_http_utils.hpp"
 #include "test_utils.hpp"
+#include "platform_utils.hpp"
 
 using namespace ovms;
