+++ src/test/embeddingsnode_test.cpp
@@ -23,6 +23,7 @@
 #include "rapidjson/document.h"
 #include "test_http_utils.hpp"
 #include "test_utils.hpp"
+#include "platform_utils.hpp"
 
 using namespace ovms;
