+++ src/test/ensemble_flow_custom_node_tests.cpp
@@ -54,6 +54,8 @@
 #include "../precision.hpp"
 #include "../stringutils.hpp"
 #include "test_utils.hpp"
+#include "light_test_utils.hpp"
+#include "test_with_temp_dir.hpp"
 
 using namespace ovms;
