+++ src/test/ensemble_tests.cpp
@@ -46,7 +46,10 @@
 #include "../status.hpp"
 #include "../tensor_conversion.hpp"
 #include "../timer.hpp"
+#include "platform_utils.hpp"
 #include "test_utils.hpp"
+#include "light_test_utils.hpp"
+#include "test_with_temp_dir.hpp"
 
 using namespace ovms;
 using namespace tensorflow;
