+++ src/test/llm/llmtemplate_test.cpp
@@ -46,6 +46,8 @@
 
 #include "../test_http_utils.hpp"
 #include "../test_utils.hpp"
+#include "src/test/light_test_utils.hpp"
+#include "../test_with_temp_dir.hpp"
 
 using namespace ovms;
