+++ src/test/llm/max_model_length_test.cpp
@@ -21,6 +21,8 @@
 
 #include "../../llm/servable_initializer.hpp"
 #include "../test_utils.hpp"
+#include "src/test/light_test_utils.hpp"
+#include "../test_with_temp_dir.hpp"
 
 using namespace ovms;
