+++ src/test/mediapipe_framework_test.cpp
@@ -42,6 +42,8 @@
 #include "../tfs_frontend/tfs_utils.hpp"
 #include "c_api_test_utils.hpp"
 #include "test_utils.hpp"
+#include "platform_utils.hpp"
+#include "test_with_temp_dir.hpp"
 
 using namespace ovms;
