+++ src/test/mediapipe_validation_test.cpp
@@ -27,6 +27,7 @@
 #include "../precision.hpp"
 #include "../server.hpp"
 #include "test_utils.hpp"
+#include "platform_utils.hpp"
 
 using namespace ovms;
