+++ src/test/metric_config_test.cpp
@@ -26,6 +26,9 @@
 #include "../modelconfig.hpp"
 #include "../modelinstance.hpp"
 #include "test_utils.hpp"
+#include "light_test_utils.hpp"
+#include "platform_utils.hpp"
+#include "test_with_temp_dir.hpp"
 
 using namespace ovms;
