+++ src/test/metrics_flow_test.cpp
@@ -38,8 +38,11 @@
 #include "../servablemanagermodule.hpp"
 #include "../server.hpp"
 #include "../shape.hpp"
+#include "platform_utils.hpp"
 #include "test_http_utils.hpp"
 #include "test_utils.hpp"
+#include "light_test_utils.hpp"
+#include "test_with_temp_dir.hpp"
 
 using namespace ovms;
