+++ src/test/modelinstance_test.cpp
@@ -26,6 +26,7 @@
 #include "../modelinstanceunloadguard.hpp"
 #include "gpuenvironment.hpp"
 #include "test_utils.hpp"
+#include "test_with_temp_dir.hpp"
 
 using testing::Return;
