+++ src/test/modelmanager_test.cpp
@@ -33,6 +33,9 @@
 #include "absl/synchronization/notification.h"
 #include "mockmodelinstancechangingstates.hpp"
 #include "test_utils.hpp"
+#include "light_test_utils.hpp"
+#include "platform_utils.hpp"
+#include "test_with_temp_dir.hpp"
 
 using testing::_;
 using testing::ContainerEq;
