+++ src/test/multipart_calculator_test.cpp
@@ -23,6 +23,7 @@
 #include "../server.hpp"
 #include "test_http_utils.hpp"
 #include "test_utils.hpp"
+#include "platform_utils.hpp"
 
 class MultiPartCalculatorTest : public ::testing::Test {
 protected:
