+++ src/test/node_library_manager_test.cpp
@@ -17,7 +17,10 @@
 #include <gtest/gtest.h>
 
 #include "../dags/custom_node_library_manager.hpp"
+#include "platform_utils.hpp"
 #include "test_utils.hpp"
+#include "light_test_utils.hpp"
+#include "test_with_temp_dir.hpp"
 
 using namespace ovms;
