+++ src/test/openvino_remote_tensors_tests.cpp
@@ -33,7 +33,8 @@
 #include "../status.hpp"
 #include "c_api_test_utils.hpp"
 #include "gpuenvironment.hpp"
-#include "test_utils.hpp"
+#include "light_test_utils.hpp"
+#include "test_with_temp_dir.hpp"
 
 using namespace ov;
