+++ src/test/pull_gguf_hf_model_test.cpp
@@ -23,6 +23,7 @@
 #include <git2.h>
 #include "env_guard.hpp"
 #include "test_utils.hpp"
+#include "test_with_temp_dir.hpp"
 #include "gguf_environment.hpp"
 #include "../filesystem.hpp"
 #include "../localfilesystem.hpp"
