+++ src/test/stateful_config_test.cpp
@@ -20,6 +20,9 @@
 #include "../modelconfig.hpp"
 #include "../modelinstance.hpp"
 #include "test_utils.hpp"
+#include "light_test_utils.hpp"
+#include "platform_utils.hpp"
+#include "test_with_temp_dir.hpp"
 
 using namespace ovms;
