+++ src/test/stateful_modelinstance_test.cpp
@@ -43,6 +43,9 @@
 #include "../timer.hpp"
 #include "stateful_test_utils.hpp"
 #include "test_utils.hpp"
+#include "light_test_utils.hpp"
+#include "platform_utils.hpp"
+#include "test_with_temp_dir.hpp"
 
 using testing::Return;
