+++ src/test/streaming_test.cpp
@@ -29,6 +29,7 @@
 #include "../status.hpp"
 #include "../stringutils.hpp"
 #include "mediapipe/framework/port/integral_types.h"
+#include "platform_utils.hpp"
 #include "test_utils.hpp"
 
 #if (PYTHON_DISABLE == 0)
