+++ src/test/tensor_conversion_test.cpp
@@ -28,6 +28,7 @@
 #include "../tensor_conversion.hpp"
 #include "opencv2/opencv.hpp"
 #include "test_utils.hpp"
+#include "platform_utils.hpp"
 
 using namespace ovms;
