+++ src/test/test_utils.cpp
@@ -14,6 +14,8 @@
 // limitations under the License.
 //*****************************************************************************
 #include "test_utils.hpp"
+#include "light_test_utils.hpp"
+#include "platform_utils.hpp"
 
 #include <algorithm>
 #include <chrono>
@@ -137,25 +139,6 @@ void waitForOVMSResourcesCleanup(ovms::ModelManager& manager) {
     const uint32_t waitTime = WAIT_MULTIPLIER_FACTOR * manager.getResourcesCleanupIntervalMillisec();
     SPDLOG_DEBUG("waitForOVMSResourcesCleanup {} ms", waitTime);
     std::this_thread::sleep_for(std::chrono::milliseconds(waitTime));
-}
-
-bool createConfigFileWithContent(const std::string& content, std::string filename) {
-    std::ofstream configFile{filename};
-    // Check if the file was successfully opened
-    if (!configFile.is_open()) {
-        SPDLOG_ERROR("Failed to open file: {}", filename);
-        throw std::runtime_error("Failed to open file: " + filename);
-    }
-    SPDLOG_INFO("Creating config file: {}\n with content:\n{}", filename, content);
-    configFile << content << std::endl;
-    configFile.close();
-    if (configFile.fail()) {
-        SPDLOG_INFO("Closing configFile failed");
-        return false;
-    } else {
-        SPDLOG_INFO("Closing configFile succeed");
-    }
-    return true;
 }
 
 ovms::tensor_map_t prepareTensors(
@@ -648,23 +631,6 @@ std::string* findKFSInferInputTensorContentInRawInputs(::KFSRequest& request, co
     return content;
 }
 
-std::string GetFileContents(const std::string& filePath) {
-    if (!std::filesystem::exists(filePath)) {
-        std::cout << "File does not exist: " << filePath << std::endl;
-        throw std::runtime_error("Failed to open file: " + filePath);
-    }
-
-    std::ifstream file(filePath, std::ios::in | std::ios::binary);
-    if (!file.is_open()) {
-        std::cout << "File could not be opened: " << filePath << std::endl;
-        throw std::runtime_error("Failed to open file: " + filePath);
-    }
-
-    std::string content{std::istreambuf_iterator<char>(file), std::istreambuf_iterator<char>()};
-    file.close();
-    return content;
-}
-
 void prepareCAPIInferInputTensor(ovms::InferenceRequest& request, const std::string& name, const std::tuple<ovms::signed_shape_t, const ovms::Precision>& inputInfo,
     const std::vector<float>& data, uint32_t decrementBufferSize, OVMS_BufferType bufferType, std::optional<uint32_t> deviceId) {
     auto [shape, type] = inputInfo;
@@ -1036,32 +1002,4 @@ void adjustConfigToAllowModelFileRemovalWhenLoaded(ovms::ModelConfig& modelConfi
     modelConfig.setPluginConfig(ovms::plugin_config_t({{"ENABLE_MMAP", "NO"}}));
 #endif
     // on linux we can remove files from disk even if mmap is enabled
-}
-std::string dirTree(const std::string& path, const std::string& indent) {
-    if (!std::filesystem::exists(path)) {
-        SPDLOG_ERROR("Path does not exist: {}", path);
-        return "NON_EXISTENT_PATH";
-    }
-    std::stringstream tree;
-    // if is directory, add to stream its name followed by "/"
-    // if is file, add to stream its name
-
-    tree << indent;
-    if (!indent.empty()) {
-        tree << "|-- ";
-    }
-
-    tree << std::filesystem::path(path).filename().string();
-    if (std::filesystem::is_directory(path)) {
-        tree << "/";
-    }
-    tree << std::endl;
-    if (!std::filesystem::is_directory(path)) {
-        return tree.str();
-    }
-    for (const auto& entry : std::filesystem::directory_iterator(path)) {
-        std::string passDownIndent = indent.empty() ? "|   " : (indent + "    ");
-        tree << dirTree(entry.path().string(), passDownIndent);
-    }
-    return tree.str();
 }
